TINKERPOP-1698 Refactored gryo serialization tests
Did a better job parameterizing tests for re-use.
